\name{barley}
\alias{barley}
\title{
A barley split plot field trial
}
\description{
An experiment involving barley and manure (nitrogen) was conducted in 6
blocks of 3 whole plots.  
}
\details{
Each whole plot was divided into 4 subplots.  Three varieties of barley were
used in the experiment with one variety being sown in each whole plot,
while the four levels of manure (0, 0.01, 0.02, and 0.04 tons per acre)
were used, one level in each of the four subplots of each whole plot.
}
\value{
The \code{barley} object is a data frame of four components:
\item{B}{
The block, as a factor with levels \code{B1}, \code{B2} \dots{} \code{B6}
}
\item{V}{
Barley varieties, as a factor with levels \code{V1} \code{V2} \code{V3}
}
\item{N}{
Nitrogen applied, in hundredths of tons per acre.
}
\item{Y}{
Yield or barley
}}
\section{CATEGORY}{
Multistratum analysis of variance.
}
\source{
Unknown. Traditional data.
}
\keyword{datasets}
% Converted by Sd2Rd version 0.3-3.
