\name{beaujolais}
\alias{beaujolais}
\title{
The Beaujolais Quality Data
}
\description{
Quality measurements for some identified samples of young Beaujolais.
}
\details{
The row names data frame \code{beaujolais} are the labels; the four
components \code{OQ}, \code{AC}, \code{pH} and \code{TSO} are the measurements.
}
\section{CATEGORY}{
Multivariate analysis
}
\source{
Quoted in Weekes: "A Genstat Primer"


Extracted from Table 1 in M. G. Jackson et al: Red wine quality:
correlations between colour, aroma and flavour and pigment and other
parameters of young Beaujolais, "Journal of Science of Food and
Agriculture", Vol 29, 715--727, 1978.
}
\keyword{datasets}
% Converted by Sd2Rd version 0.3-3.
