\name{chlorine}
\alias{chlorine}
\title{
The Chlorine availability data.
}
\description{
These industrial chemical data show the amount of chlorine
available in a certain product at various times of testing after manufacture.


The data frame has two components: \code{Time} and \code{Chlor}.
}
\section{CATEGORY}{
Nonlinear regression
}
\source{
Draper \& Smith, "Applied Regression Analysis", adapted.
}
\keyword{datasets}
% Converted by Sd2Rd version 0.3-3.
