\name{cloud}
\alias{cloud}
\title{
The Cloud Point Data
}
\description{
The cloud point of a liquid is a measure of the degree of crystalization
in a stock that can be measured by the refractive index.  It has been
suggested that the percentage of I8 in the base stock is an excellent
predictor of cloud point using the second or third order polynomial.
}
\details{
The data in the \code{cloud} data frame were collected on stocks with known
percentage of I8. The components are as follows
\item{Ispc}{
The known percentage of I8.
}
\item{Cloudpt}{
The cloud point measurement.
}}
\section{CATEGORY}{
Polynomial regression. Simple plots.
}
\source{
Draper \& Smith, "Applied Regression Analysis", page 162.
}
\keyword{datasets}
% Converted by Sd2Rd version 0.3-3.
