\name{genotype}
\alias{genotype}
\title{
The Rat Genotype Data
}
\description{
Data from a foster feeding experiment with rat mothers and litters of four
different genotypes: A, B, I and J.  The measurement is the litter weight
gain after a trial feeding period.
}
\value{
The data frame has the following components:
\item{Litter}{
The genotype of the litter 
}
\item{Mother}{
The genotype of the mother
}
\item{Wt}{
Litter weight
}}
\section{CATEGORY}{
Unbalanced double classification.
}
\source{
Quoted in Scheffe, H.: "The Analysis of Variance"
}
\keyword{datasets}
% Converted by Sd2Rd version 0.3-3.
