\name{house}
\alias{house}
\title{
The Tuggeranong House Price Data
}
\description{
Before buying a house in Tuggeranong in February, 1987, a cautious
potential householder collected some data on houses on the market.  
}
\value{
The
data for 20 such houses is available in the data frame \code{house}, whose components are as follows
\item{Price}{
Cost of house, in tens of thousands of dollars
}
\item{Floor}{
Total floor area
}
\item{Block}{
Area of land
}
\item{Rooms}{
Number of main rooms
}
\item{Age}{
Age of house, in years
}
\item{Cent.Heat}{
Indicates whether the house was centrally heated.
}}
\section{CATEGORY}{
Multiple regression, coplots.
}
\source{
Dr Ray Correll, personal communication.
}
\keyword{datasets}
% Converted by Sd2Rd version 0.3-3.
