\name{iowheat}
\alias{iowheat}
\title{
The Iowa Wheat Yield Data
}
\description{
The data gives the pre-season and three growing months' precipitation, the
mean temperatures for the three growing months and harvest month, the year,
and the yield of wheat for the USA state of Iowa, for the years 1930--1962.
}
\value{
The data frame has the following components:
\item{Year}{
Year of measurements
}
\item{Rain0}{
Pre-season rainfall
}
\item{Temp1}{
Mean temperature for first growing month
}
\item{Rain1}{
Rainfall for first growing month
}
\item{Temp2}{
Mean temperature for second growing month
}
\item{Rain2}{
Rainfall for first growing month
}
\item{Temp3}{
Mean temperature for third growing month
}
\item{Rain3}{
Rainfall for first growing month
}
\item{Temp4}{
Mean temperature for harvest month
}
\item{Yield}{
Yield of wheat in Iowa for given year
}}
\section{CATEGORY}{
Multiple regression; diagnostics.
}
\source{
CAED Report, 1964.  Quoted in Draper \& Smith.
}
\keyword{datasets}
% Converted by Sd2Rd version 0.3-3.
