\name{jellyfish}
\alias{jellyfish}
\title{
The Jellyfish Data
}
\description{
Two samples of jellyfish, from Danger Island and Salamander Bay
respectively, were measured for length and width.
}
\value{
The data frame has the following components:
\item{Sample}{
Location of jellyfish, as a factor: \code{D} for Danger Island, and 
\code{S} for Salamander Bay.
}
\item{Width}{
Width of jellyfish
}
\item{Length}{
Length of jellyfish
}}
\section{CATEGORY}{
Bivariate, two sample data.
}
\source{
"Interactive Statistics", Ed. Don McNeil.
}
\keyword{datasets}
% Converted by Sd2Rd version 0.3-3.
