\name{kalythos}
\alias{kalythos}
\title{
The Kalythos Blindness Data
}
\description{
On the Greek island of Kalythos the male inhabitants suffer from a
congenital eye disease, the effects of which become more marked with
increasing age.  Samples of islander males of various ages  were tested for
blindness and the results recorded.
}
\value{
The data frame contains the following components:
\item{Age}{
Age of subjects
}
\item{N}{
Number of subjects at this age
}
\item{Blind}{
Number of these subjects which are actually blind.
}}
\section{CATEGORY}{
Generalized linear modelling
}
\source{
S. D. Silvey: "Statistical Inference".  (Fictitious?)
}
\keyword{datasets}
% Converted by Sd2Rd version 0.3-3.
