\name{oil}
\alias{oil}
\title{
The Gasoline Yield Data
}
\description{
The data gives the gasoline yield as a percent of crude oil, and
four independent variables which may influence yield. 
}
\details{
The data come as 10 separate samples, and within each sample the values
for the independent variables are constant.
}
\value{
The data frame contains the following components:
\item{Sample}{
The sample number, from 1 to 10
}
\item{API}{
The crude oil gravity, in degrees API
}
\item{Vapour}{
The crude oil vapour pressure
}
\item{ASTM}{
The crude oil 10\% point
}
\item{Endpt}{
The gasoline end point
}
\item{Yield}{
Percentage of oil recovered
}}
\section{CATEGORY}{
Analysis of variance, covariance, and multiple regression. Modern regression.
}
\source{
"Estimate gasoline yields from crudes",
by Nilon H.P rater, "Petroleum Refiner", Vol 35, No. 5.
}
\keyword{datasets}
% Converted by Sd2Rd version 0.3-3.
