\name{painters}
\alias{painters}
\title{
The Painter's Data of de Piles
}
\description{
The subjective assessment, on a 0 to 20 integer scale, of 54
classical painters.  The painters were assessed on four characteristics:
composition, drawing, colour and expression.  The data is due to the
Eighteenth century art critic, de Piles.
}
\value{
The row names of the data frame are the painters. The components are:
\item{Composition}{
Composition score
}
\item{Drawing}{
Drawing score
}
\item{Colour}{
Colour score
}
\item{Expression}{
Expression score
}
\item{School}{
The school to which a painter belongs, as indicated by a factor level
code as follows:
A:Renaissance; 
B:Mannerist;
C:Seicento;
D:Venetian;
E:Lombard;
F:Sixteenth Century;
G:Seventeenth Century;
H:French.
}}
\section{CATEGORY}{
Multivariate Analysis: Discriminant Analysis.
}
\source{
Weekes: A Genstat Primer.
}
\keyword{datasets}
% Converted by Sd2Rd version 0.3-3.
