\name{pottery}
\alias{pottery}
\title{
Archaelogical Pottery Data
}
\description{
The data arise from a chemical analysis of 26 samples of pottery found
at Romano-British kiln sites in Wales, Gwent and the New Forest.  The
variables describe the composition, in terms of the metals aluminium,
iron, magnesium, calcium and sodium, and are expressed as percentages
of the oxides of the metals.
}
\value{
The data frame contains the following components:
\item{Site}{
The site, coded as follows: L:Llanederyn; C:Caldicot; I:Island Thorns;
A:Ashley Rails
}
\item{Al}{
Aluminiun content
}
\item{Fe}{
Iron content
}
\item{Mg}{
Magnesium content
}
\item{Ca}{
Calcium content
}
\item{Na}{
Sodium content
}}
\section{CATEGORY}{
Multivariate analysis
}
\source{
Tubb, A. et al. "Archaeometry", Vol 22, pp 153--171, (1980)
}
\keyword{datasets}
% Converted by Sd2Rd version 0.3-3.
