\name{rumford}
\alias{rumford}
\title{
Count Rumford's Friction Data
}
\description{
Data on the amount of heat generated by friction was obtained by Lord
Rumford in 1798.
}
\details{
A bore was fitted into a stationary cylinder and pressed
against the bottom by a screw.  The bore was turned by a team of horses
for 30 minutes, after which Lord Rumford ``suffered the thermometer to
remain in its place nearly three quarters of an hour, observing and noting
down, at small intervals of time, the temperature indicated by it''.
\value{
\item{Time}{
Time in minutes.
}
\item{Temp}{
Temperature in degrees Farenheit.
}}
}
\source{
Bates \& Watts, "Nonlinear Regression Analysis".
}
\keyword{datasets}
% Converted by Sd2Rd version 0.3-3.
