\name{sawheat}
\alias{sawheat}
\title{
Yorke Penninsula Wheat Data
}
\description{
The annual yield of wheat in a marginal wheat growing district on the Yorke
Peninsula, South Australia, together with the rainfall for the three
growing months, for the years 1931--1955.  The year itself is potentially a
surrogate predictor to allow for improvements in varieties and farm
practice.
}
\value{
The data frame contains the following components:
\item{Year}{
Year of measurement
}
\item{Rain0}{
Rainfall for first growing month, in inches.
}
\item{Rain1}{
Rainfall for second growing month, in inches.
}
\item{Rain2}{
Rainfall for third growing month, in inches.
}
\item{Yield}{
Yield of wheat, in bushels per acre.
}}
\section{CATEGORY}{
Multiple regression
}
\source{
K. W. Morris (private communication)
}
\keyword{datasets}
% Converted by Sd2Rd version 0.3-3.
