\name{steam}
\alias{steam}
\title{
The Saturated Steam Pressure Data
}
\description{
Temperature and pressure (Pascals) in a
saturated steam driven experimental device.
}
\value{
The data frame contains the following components:
\item{Temp}{
Temperature, in degrees Celsius
}
\item{Press}{
Presure, in Pascals
}}
\section{CATEGORY}{
Nonlinear regression
}
\source{
Quoted in Draper \& Smith: "Applied Regression Analysis"
}
\keyword{datasets}
% Converted by Sd2Rd version 0.3-3.
